+++ src/utils/utils.ts
@@ -18,28 +18,3 @@ export const toDateTimeSeconds = (date: Date) => {
 };
 
 export const dateReadable = (timestamp: number) => toDateString(new Date(timestamp * 1000));
-
-export const defaultsTo = <T>(value: T | null | undefined, defaultValue: T) => (notEmpty<T>(value) ? value : defaultValue);
-
-export const defaultsToObject = <T extends object>(value?: T | null): T => defaultsTo(value, {} as T);
-
-export const defaultsToArray = <T>(value?: T[] | null): T[] => defaultsTo(value, []);
-
-export const prop = (property: string, object?: { [key: string]: any }) => (object && Object.prototype.hasOwnProperty.call(object, property) ? object[property] : undefined);
-export const propFc = (property: string) => (object?: { [key: string]: any }) => prop(property, object);
-
-export const head = <T>(arr: T[]): T | undefined => (arr ? arr[0] : undefined);
-
-export const has = <T extends object>(property: keyof T) => (obj: T) => !!obj[property];
-
-export const path = <T>(object?: object, ...bits: any[]): T | undefined => {
-  const [property, ...rest] = bits;
-  if (prop(property, object)) {
-    if (head(rest)) {
-      return path<T>(prop(property, object), ...rest);
-    }
-    return prop(property, object);
-  }
-  return undefined;
-};
-export const pathFc = <TResult>(bits: string[]) => (object: object): TResult | undefined => path<TResult>(object, ...bits);